Deleted and Split QuickCalcController. Combined UnitSpec and OneAppPerSuite in AppUnitGenerator
