+++ test/uk/gov/hmrc/payetaxcalculatorfrontend/controllers/QuickCalcControllerSpec/SubmitTaxCodeGenerator.scala
@@ -1,0 +1,154 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.payetaxcalculatorfrontend.controllers.QuickCalcControllerSpec
+
+import org.jsoup.Jsoup
+import play.api.test.Helpers._
+import uk.gov.hmrc.payetaxcalculatorfrontend.controllers.QuickCalcController
+import uk.gov.hmrc.payetaxcalculatorfrontend.model.UserTaxCode
+import uk.gov.hmrc.payetaxcalculatorfrontend.setup.AppUnitGenerator
+import uk.gov.hmrc.payetaxcalculatorfrontend.setup.QuickCalcCacheSetup._
+import uk.gov.hmrc.play.http.SessionKeys
+
+class SubmitTaxCodeGenerator extends AppUnitGenerator {
+
+  "Submit Tax Code Form" should {
+    "return 400, current list of aggregate data and an error message for invalid Tax Code" in {
+      val controller = new QuickCalcController(messages.messages, cacheReturnTaxCode)
+      val formTax = UserTaxCode.form.fill(UserTaxCode(true, Some("110")))
+      val action = await(csrfAddToken(controller.submitTaxCodeForm()))
+      val result = action(request.withFormUrlEncodedBody(formTax.data.toSeq: _*)).withSession(request.session + (SessionKeys.sessionId -> "test-tax"))
+      val status = result.header.status
+      val responseBody = contentAsString(result)
+      val parseHtml = Jsoup.parse(responseBody)
+
+      val expectedErrorMessage = "The tax code you have entered is not valid - it must end with the letter L, ‘M, ‘N, or T"
+      val expectedTableSize = 1 + aggregateListOnlyTaxCode.size // include header
+
+      val actualTableSize = parseHtml.getElementsByTag("tr").size()
+      val actualErrorMessage = parseHtml.getElementsByClass("error-notification").text()
+
+      status shouldBe 400
+      actualErrorMessage shouldBe expectedSuffixTaxCodeErrorMessage
+      actualTableSize shouldBe expectedTableSize
+    }
+
+    "return 400, empty list of aggregate data and an error message for invalid Tax Code" in {
+      val controller = new QuickCalcController(messages.messages, cacheEmpty)
+      val formTax = UserTaxCode.form.fill(UserTaxCode(true, Some("110")))
+      val action = await(csrfAddToken(controller.submitTaxCodeForm()))
+      val result = action(request.withFormUrlEncodedBody(formTax.data.toSeq: _*)).withSession(request.session + (SessionKeys.sessionId -> "test-tax"))
+      val status = result.header.status
+      val responseBody = contentAsString(result)
+      val parseHtml = Jsoup.parse(responseBody)
+
+      val expectedErrorMessage = "The tax code you have entered is not valid - it must end with the letter L, ‘M, ‘N, or T"
+
+      val actualTableSize = parseHtml.getElementsByTag("tr").size()
+      val actualErrorMessage = parseHtml.getElementsByClass("error-notification").text()
+
+      status shouldBe 400
+      actualErrorMessage shouldBe expectedSuffixTaxCodeErrorMessage
+      actualTableSize shouldBe 0
+    }
+
+    "return 400, empty list of aggregate data and an error message when user selects \"Yes\" but did not enter Tax code" in {
+      val controller = new QuickCalcController(messages.messages, cacheEmpty)
+      val formTax = UserTaxCode.form.fill(UserTaxCode(true, None))
+      val action = await(csrfAddToken(controller.submitTaxCodeForm()))
+      val result = action(request.withFormUrlEncodedBody(formTax.data.toSeq: _*)).withSession(request.session + (SessionKeys.sessionId -> "test-tax"))
+      val status = result.header.status
+      val responseBody = contentAsString(result)
+      val parseHtml = Jsoup.parse(responseBody)
+
+      val actualTableSize = parseHtml.getElementsByTag("tr").size()
+      val actualErrorMessage = parseHtml.getElementsByClass("error-notification").text()
+
+      status shouldBe 400
+      actualErrorMessage shouldBe expectedInvalidTaxCodeErrorMessage
+      actualTableSize shouldBe 0
+    }
+
+    "return 400 when Tax Code Form submission is empty" in {
+      val controller = new QuickCalcController(messages.messages, cacheEmpty)
+      val formTax = UserTaxCode.form
+      val postAction = await(csrfAddToken(controller.submitTaxCodeForm()))
+
+      val postResult = postAction(request
+        .withFormUrlEncodedBody(formTax.data.toSeq: _*))
+        .withSession(SessionKeys.sessionId -> "test-tax")
+
+      val status = postResult.header.status
+      status shouldBe 400
+    }
+
+    "return 303, when Tax Code Form submission, current list of aggregate and redirect to Is Over State Pension Page" in {
+      val controller = new QuickCalcController(messages.messages, cacheReturnTaxCodeAndIsOverStatePension)
+      val formTax = UserTaxCode.form.fill(UserTaxCode(true, Some("K425")))
+      val postAction = await(csrfAddToken(controller.submitTaxCodeForm()))
+
+      val postResult = postAction(request
+        .withFormUrlEncodedBody(formTax.data.toSeq: _*))
+        .withSession(SessionKeys.sessionId -> "test-tax")
+
+      val status = postResult.header.status
+      val actualRedirectUri = redirectLocation(postResult).get
+
+      val expectedRedirectUri = "/paye-tax-calculator/quick-calculation/age"
+
+      status shouldBe 303
+      actualRedirectUri shouldBe expectedRedirectUri
+    }
+
+    "return 303, when Tax Code Form submission, new list of aggregate and redirect to Is Over State Pension Page" in {
+      val controller = new QuickCalcController(messages.messages, cacheEmpty)
+      val formTax = UserTaxCode.form.fill(UserTaxCode(true, Some("K425")))
+      val postAction = await(csrfAddToken(controller.submitTaxCodeForm()))
+
+      val postResult = postAction(request
+        .withFormUrlEncodedBody(formTax.data.toSeq: _*))
+        .withSession(SessionKeys.sessionId -> "test-tax")
+
+      val status = postResult.header.status
+      val actualRedirectUri = redirectLocation(postResult).get
+
+      val expectedRedirectUri = "/paye-tax-calculator/quick-calculation/age"
+
+      status shouldBe 303
+      actualRedirectUri shouldBe expectedRedirectUri
+    }
+
+    "return 303, when Tax Code form submission, the complete list of aggregate data and redirect to Summary Result Page" in {
+      val controller = new QuickCalcController(messages.messages, cacheReturnTaxCodeIsOverStatePensionAndSalary)
+      val formTax = UserTaxCode.form.fill(UserTaxCode(true, Some("K425")))
+      val postAction = await(csrfAddToken(controller.submitTaxCodeForm()))
+
+      val postResult = postAction(request
+        .withFormUrlEncodedBody(formTax.data.toSeq: _*))
+        .withSession(SessionKeys.sessionId -> "test-tax")
+
+      val status = postResult.header.status
+      val actualRedirectUri = redirectLocation(postResult).get
+
+      val expectedRedirectUri = "/paye-tax-calculator/quick-calculation/summary-result"
+
+      status shouldBe 303
+      actualRedirectUri shouldBe expectedRedirectUri
+    }
+  }
+
+}
